==> notebooks/10_exploratory_data_analysis.ipynb <==
# Notebook Initialization

# %load nb_init.py

from pathlib import Path
import pandas as pd

base_dir = Path.cwd().parent
config_dir = base_dir / "config"
data_dir = base_dir / "data"
input_dir = data_dir / "COVID19"
preprocessed_dir = data_dir / "preprocessed"
output_dir = data_dir / "output"

metadata_file = input_dir / "metadata.csv"
labels_file = input_dir / "unzip_filenames.csv"
preprocessed_labels_file = preprocessed_dir / "labels.parquet"

from pyspark.sql import SparkSession
from pyspark.sql import functions as F

spark = SparkSession.builder.getOrCreate()

config_file = config_dir / "tfg.conf"

from pyhocon import ConfigFactory
config = ConfigFactory.parse_file(config_file).tfg


config

spark

csv_options = config.get("eda.csv_options").as_plain_ordered_dict()
csv_options

# Demographics
## Metadata file

!head -n2 {metadata_file}

metadata = spark.read\
    .options(**csv_options)\
    .csv(str(metadata_file))

metadata.show(5)

metadata.printSchema()

metadata_exprs = [
    "patient_id AS patient_id",
    "scan_id AS scan_id",
    "Age AS age",
    "`Sex(Male1/Female2)` AS gender",
    "Critical_illness AS critical_illness",
    "Liver_function AS liver_function",
    "Lung_function AS lung_function",
    "`Progression (Days)` AS progression_days",
]

metadata = metadata.selectExpr(*metadata_exprs)

metadata.show(5)

metadata.count()

## Labels

labels = spark.read.options(**csv_options).csv(str(labels_file))

labels.show(5)

num_patients = labels.select("patient_id").distinct().count()
num_scans = labels.select("patient_id", "scan_id").distinct().count()
print(f"{num_patients = }")
print(f"{num_scans = }")

labels.printSchema()

labels_expr = [
    "patient_id AS patient_id",
    "scan_id AS scan_id",
    "n_slice AS n_slice",
    "label",
]

labels = labels.selectExpr(*labels_expr)

labels.show(5)

total_labels = labels.count()
total_labels

## Analyse number of slices per CT

We already know that we'll be dividing the CTs into clips of a fixed size, where each clip will contain a specific number of images N.

labels.describe("n_slice").show()

n_values = [50, 70, 100, 150, 200, 300]
for n in n_values:
    labels_in_n = labels.filter(f"n_slice <= {n}").count()
    print(f"There are {labels_in_n} / {total_labels} ({100 * labels_in_n / total_labels:.2f}%) CTs with <= {n} images")

Check how many "clips" we'd get for each value of N. i.e. if we have a CT with 100 slices, and we chose N = 50, we'd get 2 clips.

To be checked if it would be better to have overlapping clips instead, e.g. for 100 slices and N=50, have a clip for images 0-50, another one for images 25-75, and another one for 50-100.

label_clips = labels

for n in n_values:
    label_clips = label_clips\
        .withColumn(f"num_clips_{n}", F.ceil(labels.n_slice / n))

label_clips.show()

agg_expressions = []

for n in n_values:
    agg_expressions.append(F.sum(f"num_clips_{n}").alias(f"num_clips_{n}"))

label_clips\
    .groupBy()\
    .agg(*agg_expressions)\
    .show()

## Conclusion

We'll pick a value of N=70. This value is relatively small, gives us a decent amount of clips, and around 50% of all the CTs can be fit into a single clip.

labels = labels.withColumn("num_clips", F.ceil(labels.n_slice / 70))

labels.show()

labels.coalesce(1).write.parquet(str(preprocessed_labels_file))

labels = spark.read.parquet(str(preprocessed_labels_file))

## Check overlap between labels and metadata

Do we have demographics for patients for which we have data?

total_labels = labels.count()
total_labels_with_demo = labels.join(metadata, ["patient_id"], "left_semi").count()

print(f"We have demographics for {total_labels_with_demo} / {total_labels} observations ({100 * total_labels_with_demo / total_labels:.2f}%)")

labels_with_metadata = labels.join(metadata, ["patient_id"], "left_semi")

labels_with_metadata.count()

labels_with_metadata.groupBy("label").count().show()

Is there also overlap on patient_id AND scan_id level?

labels.join(metadata, ["patient_id", "scan_id"], "left_semi").groupBy("label").count().show()

labels_pd = labels.toPandas()

labels_pd

labels.select("patient_id").count(), labels.select("patient_id").distinct().count()

## Do any patient_ids have more than 1 label?

from pyspark.sql import functions as F

labels\
    .groupBy("patient_id")\
    .agg(F.countDistinct("label").alias("num_labels"))\
    .filter("num_labels > 1")\
    .count()

## Check number labels with / without unique patient ids

total_labels = labels.count()
total_slices = labels.selectExpr("sum(n_slice) AS total").first().total

total_labels, total_slices

labels\
    .groupBy("label")\
    .agg(
        F.count("*").alias("count"),
        F.sum("n_slice").alias("n_slice")
    )\
    .withColumn("count_pct", F.expr(f"ROUND(count / {total_labels}, 4)"))\
    .withColumn("slice_pct", F.expr(f"ROUND(n_slice / {total_slices}, 4)"))\
    .orderBy("label")\
    .show()

unique_patients = labels.select("patient_id").distinct().count()
labels\
    .dropDuplicates(["patient_id"])\
    .groupBy("label")\
    .count()\
    .withColumn("pct", F.expr(f"ROUND(count / {unique_patients}, 4)"))\
    .orderBy("label")\
    .show()

## Check labels with metadata only

total_labels_with_metadata = labels_with_metadata.count()
total_slices_with_metadata = labels_with_metadata.selectExpr("sum(n_slice) AS total").first().total

print(f"Total labels with metadata: {total_labels_with_metadata}")
print(f"Total slices with metadata: {total_slices_with_metadata}")

labels_with_metadata\
    .groupBy("label")\
    .agg(
        F.count("*").alias("count"),
        F.sum("n_slice").alias("n_slice")
    )\
    .withColumn("count_pct", F.expr(f"ROUND(count / {total_labels_with_metadata}, 4)"))\
    .withColumn("slice_pct", F.expr(f"ROUND(n_slice / {total_slices_with_metadata}, 4)"))\
    .orderBy("label")\
    .show()

unique_patients_with_metadata = labels_with_metadata.select("patient_id").distinct().count()

print(f"There are {unique_patients_with_metadata} unique patients with metadata")
labels_with_metadata\
    .dropDuplicates(["patient_id"])\
    .groupBy("label")\
    .count()\
    .withColumn("pct", F.expr(f"ROUND(count / {unique_patients_with_metadata}, 4)"))\
    .orderBy("label")\
    .show()

## Conclusion

There is almost no metadata for patients with NCP (there's metadata only for 13). It could be usable if we only want to consider e.g. CP VS Normal, but won't be useful for NCP.

import sweetviz as sv

report = sv.analyze(labels_pd)

report.show_html(filepath=base_dir / "docs" / "1_eda_labels.html")

# Train / test split
