==> notebooks/20_train_val_test_splits.ipynb <==
# Notebook Initialization

# %load nb_init.py

from pathlib import Path
import pandas as pd

base_dir = Path.cwd().parent
config_dir = base_dir / "config"
data_dir = base_dir / "data"
docs_dir = base_dir / "docs"
figures_dir = docs_dir / "figures"
models_dir = base_dir / "models"
logs_dir = base_dir / "logs"
images_input_dir = data_dir / "COVID19"
images_pp_dir = data_dir / "COVID19-preprocessed"
preprocessed_dir = data_dir / "preprocessed"
output_dir = data_dir / "output"

# Directories used to train the CNN (image by image) 
cnn_data_dir = data_dir / "modelling" / "cnn"
cnn_train_dir = cnn_data_dir / "train"
cnn_test_dir = cnn_data_dir / "test"

metadata_file = images_input_dir / "metadata.csv"
labels_file = images_input_dir / "unzip_filenames.csv"
preprocessed_labels_file = preprocessed_dir / "labels.parquet"

feature_extractor_model_file = models_dir / "feature_extractor.tf"

from pyspark.sql import SparkSession
from pyspark.sql import functions as F

spark = SparkSession.builder.getOrCreate()

config_file = config_dir / "tfg.conf"

from pyhocon import ConfigFactory
config = None

def load_config():
    return ConfigFactory.parse_file(config_file)

config = load_config()
    
import sys

if str(base_dir / "src") not in sys.path:
    sys.path.append(str(base_dir / "src"))

%load_ext autoreload

%autoreload 2

spark

from tfg import DataRepository

repo = DataRepository(config=config, base_data_path=data_dir)

# Load labels

labels = spark.read.parquet(str(preprocessed_labels_file))

labels.show(5)

num_labels = labels.count()
num_labels

labels\
    .groupBy("label")\
    .agg(
        F.count("*").alias("count"),
        F.round((F.count("*") / num_labels), 4).alias("pct"),
    )\
    .show()

labels.select("patient_id").distinct().count()

# Split patient ids into train / test

seed = config.get_int("tfg.seed")
test_fraction = config.get_float("tfg.training.test_fraction")
val_fraction = config.get_float("tfg.training.val_fraction")
train_fraction = 1 - test_fraction - val_fraction
split_weights = [train_fraction, val_fraction, test_fraction]
split_weights

patient_ids = labels.select("patient_id").distinct()

id_dfs = patient_ids.randomSplit(split_weights, seed=seed)
train_ids = id_dfs[0]
val_ids =  id_dfs[1]
test_ids =  id_dfs[2]

train_labels = labels.join(train_ids, ["patient_id"], "inner")
val_labels = labels.join(val_ids, ["patient_id"], "inner")
test_labels = labels.join(test_ids, ["patient_id"], "inner")

train_count = train_labels.count()
val_count = val_labels.count()
test_count = test_labels.count()

print(f"Training rows: {train_count}")
print(f"Validation rows: {val_count}")
print(f"Test rows: {test_count}")

train_labels.show(5)
val_labels.show(5)
test_labels.show(5)

print("Train label counts and percentages")
train_labels\
    .groupBy("label")\
    .agg(
        F.count("*").alias("count"),
        F.round((F.count("*") / train_count), 2).alias("pct"),
    )\
    .show()

print("Validation label counts and percentages")
val_labels\
    .groupBy("label")\
    .agg(
        F.count("*").alias("count"),
        F.round((F.count("*") / val_count), 2).alias("pct"),
    )\
    .show()

print("Test label counts and labels")
test_labels\
    .groupBy("label")\
    .agg(
        F.count("*").alias("count"),
        F.round((F.count("*") / test_count), 2).alias("pct"),
    )\
    .show()

# Create datasets for the CNN

For each of the train/validation/test sets we have, we will create a pandas dataframe, and we'll use these dataframe to create the data generators for our model.

from dataclasses import dataclass

from pyspark.sql import Row


@dataclass
class Observation:
    patient_id: int
    scan_id: int
    n_slice: int
    num_clips: int
    label: str

    @staticmethod
    def from_row(r: Row) -> "Observation":
        return Observation(
            patient_id=r.patient_id,
            scan_id=r.scan_id,
            n_slice=r.n_slice,
            num_clips=r.num_clips,
            label=r.label,
        )

from tqdm.notebook import tqdm
import pandas as pd
from pathlib import Path

def create_df(config, labels):
    raw_images = {
        "label": [],
        "file": [],
        "patient_id": [],
        "scan_id": [],
        "n_slice": [],
        "num_clips": [],
        "seq_num": [],
        "clip_num": [],
    }

    IMAGES_PER_CLIP = config.get_int("tfg.training.images_per_clip")

    for r in labels.collect():
        obs = Observation.from_row(r)
        scan_path = Path(f"{obs.label}/{obs.patient_id}/{obs.scan_id}")
        absolute_scan_path = images_input_dir / scan_path
        # different directories have different formats
        img_ext = next(absolute_scan_path.glob("*.*")).suffix

        scan_images = []
        seq_nums = []
        clip_nums = []

        seq_num = 0
        images_input_dir_str = str(images_input_dir)
        path_offset = len(images_input_dir_str) + 1
        for seq_num, img in enumerate(absolute_scan_path.glob(f"*{img_ext}")):
            img_to_append = str(img)[path_offset:]
            scan_images.append(img_to_append)
            seq_nums.append(seq_num)
            clip_nums.append(seq_num // IMAGES_PER_CLIP)
            seq_num = seq_num + 1

        raw_images["label"].extend([obs.label] * obs.n_slice)
        raw_images["file"].extend(scan_images)
        raw_images["patient_id"].extend([obs.patient_id] * obs.n_slice)
        raw_images["scan_id"].extend([obs.scan_id] * obs.n_slice)
        raw_images["n_slice"].extend([obs.n_slice] * obs.n_slice)
        raw_images["num_clips"].extend([obs.patient_id] * obs.n_slice)
        raw_images["seq_num"].extend(seq_nums)
        raw_images["clip_num"].extend(clip_nums)
        
    return pd.DataFrame(raw_images)

train_df = create_df(config, train_labels)
print(f"{train_df.shape}")
train_df

repo.save("train_df", train_df)

val_df = create_df(config, val_labels)
print(f"{val_df.shape}")
val_df

repo.save("val_df", val_df)

test_df = create_df(config, test_labels)
print(f"{test_df.shape}")
test_df

repo.save("test_df", test_df)

# Data Generators

train_df = repo.load("train_df")
train_df.head(5)

val_df = repo.load("val_df")
val_df.head(5)

from tqdm.notebook import tqdm
from tensorflow.keras.preprocessing.image import load_img

img_sizes = set()

# We know that all images in a scan have the same size, so we only need 1 image per scan,
# that's why we can remove the duplicatees based on patient id + scan id
for r in tqdm(train_df.drop_duplicates(["patient_id", "scan_id"]).iterrows()):
    img_path = images_input_dir / r[1]["file"]
    img_sizes.add(load_img(img_path).size)

img_sizes

val_img_sizes = set()

# We know that all images in a scan have the same size, so we only need 1 image per scan,
# that's why we can remove the duplicatees based on patient id + scan id
for r in tqdm(val_df.drop_duplicates(["patient_id", "scan_id"]).iterrows()):
    img_path = images_input_dir / r[1]["file"]
    val_img_sizes.add(load_img(img_path).size)

val_img_sizes
