==> notebooks/62_convert_images_into_features.ipynb <==
# Notebook Initialization

# %load nb_init.py

from pathlib import Path
import pandas as pd

base_dir = Path.cwd().parent
config_dir = base_dir / "config"
data_dir = base_dir / "data"
docs_dir = base_dir / "docs"
figures_dir = docs_dir / "figures"
models_dir = base_dir / "models"
logs_dir = base_dir / "logs"
images_input_dir = data_dir / "COVID19"
preprocessed_dir = data_dir / "preprocessed"
output_dir = data_dir / "output"

# Directories used to train the CNN (image by image) 
cnn_data_dir = data_dir / "modelling" / "cnn"
cnn_train_dir = cnn_data_dir / "train"
cnn_test_dir = cnn_data_dir / "test"

metadata_file = images_input_dir / "metadata.csv"
labels_file = images_input_dir / "unzip_filenames.csv"
preprocessed_labels_file = preprocessed_dir / "labels.parquet"

feature_extractor_model_file = models_dir / "feature_extractor.tf"

from pyspark.sql import SparkSession
from pyspark.sql import functions as F

spark = SparkSession.builder.getOrCreate()

config_file = config_dir / "tfg.conf"

from pyhocon import ConfigFactory
config = None

def load_config():
    return ConfigFactory.parse_file(config_file)

config = load_config()
    
import sys

if str(base_dir / "src") not in sys.path:
    sys.path.append(str(base_dir / "src"))

%load_ext autoreload

%autoreload 2

spark

from tfg import DataRepository

repo = DataRepository(config=config, base_data_path=data_dir)

# Load model

from tensorflow.keras.models import load_model
cnn_model_suffix = "07"
model_file = models_dir / f"feature_extractor_{cnn_model_suffix}.tf"
model = load_model(str(model_file))

model.summary()

import tensorflow as tf

new_model = tf.keras.Sequential()

model.layers[:-2]

for l in model.layers[:-1]:
    if isinstance(l, tf.python.keras.layers.core.Dropout):
        continue
    new_model.add(l)

new_model.summary()

# Load datasets

train_df = repo.load("train_df")
train_df.head(5)

It seems like the num_clips, seq_num and clip_num are not correct, we'll fix them..

images_per_clip = config.get_int("tfg.training.images_per_clip")
print(f"{images_per_clip = }")

from pyspark.sql import Window

images_per_clip = config.get_int("tfg.training.images_per_clip")
print(f"{images_per_clip = }")

scan_window = Window\
    .partitionBy("patient_id", "scan_id")\
    .orderBy("file")

def fix_df(pandas_df):
    df = spark.createDataFrame(pandas_df)

    new_df = df\
        .withColumn("seq_num", F.row_number().over(scan_window) - 1)\
        .withColumn("num_clips", F.ceil(F.col("n_slice") / images_per_clip))\
        .withColumn("clip_num", F.floor(F.col("seq_num") / images_per_clip))\
        .toPandas()

    return new_df

train_df.head()

train_df = repo.load("train_df")
train_df = fix_df(train_df)
repo.save("train_df", train_df)
train_df.head()

val_df = repo.load("val_df")
val_df = fix_df(val_df)
repo.save("val_df", val_df)
val_df.head()

test_df = repo.load("test_df")
test_df = fix_df(test_df)
repo.save("test_df", test_df)
test_df.head()

# Convert images into features

img_size = 64#config.get_int("tfg.training.img_size")
print(f"{img_size = }")

image_target_size = (img_size, img_size)
print(f"{image_target_size = }")

seed = config.get_int("tfg.seed")
print(f"{seed = }")

batch_size = config.get_int("tfg.training.batch_size")
print(f"{batch_size = }")


from keras_preprocessing.image import ImageDataGenerator

test_datagen = ImageDataGenerator(
    rescale=1./255.,
)

## Train

train_df = repo.load("train_df")
train_generator = test_datagen.flow_from_dataframe(
    dataframe=train_df,
    directory=str(images_input_dir),
    x_col="file",
    y_col="label",
    batch_size=batch_size,
    seed=seed,
    shuffle=False,
    class_mode="categorical",
    target_size=image_target_size)

train_generator.class_indices

class_to_idx = train_generator.class_indices
idx_to_class = { v: k for k, v in class_to_idx.items() }

NUM_TRAIN_BATCHES = train_generator.n // train_generator.batch_size + 1

raw_train_feats = new_model.predict(train_generator, steps=NUM_TRAIN_BATCHES, verbose=1)
raw_train_feats.shape

train_df["img_features"] = raw_train_feats.tolist()

repo.save("train_features_final", train_df)

train_feats = repo.load("train_features_final")
train_feats.head(5)

# train_generator.reset()
# raw_train_probs = model.predict(train_generator, steps=NUM_TRAIN_BATCHES, verbose=1)
# raw_train_probs.shape

# train_feats["img_probs"] = raw_train_probs.tolist()
# train_feats.head(5)

# repo.save("train_features", train_feats)

## Validation

val_df = repo.load("val_df")

val_generator = test_datagen.flow_from_dataframe(
    dataframe=val_df,
    directory=str(images_input_dir),
    x_col="file",
    y_col="label",
    batch_size=batch_size,
    seed=seed,
    shuffle=False,
    class_mode="categorical",
    target_size=image_target_size)

NUM_VAL_BATCHES = val_generator.n // val_generator.batch_size + 1

raw_val_feats = new_model.predict(val_generator, steps=NUM_VAL_BATCHES, verbose=1)
raw_val_feats.shape

val_df["img_features"] = raw_val_feats.tolist()
val_df.head()

repo.save("val_features_final", val_df)

val_feats = repo.load("val_features_final")
val_feats.head(5)

# val_generator.reset()
# raw_val_probs = model.predict(val_generator, steps=NUM_VAL_BATCHES, verbose=1)
# raw_val_probs.shape

# val_feats["img_probs"] = raw_val_probs.tolist()
# val_feats.head(5)

# repo.save("val_features", val_feats)

## Test

test_df = repo.load("test_df")

test_generator = test_datagen.flow_from_dataframe(
    dataframe=test_df,
    directory=str(images_input_dir),
    x_col="file",
    y_col="label",
    batch_size=batch_size,
    seed=seed,
    shuffle=False,
    class_mode="categorical",
    target_size=image_target_size)

NUM_TEST_BATCHES = test_generator.n // test_generator.batch_size + 1

raw_test_feats = new_model.predict(test_generator, steps=NUM_TEST_BATCHES, verbose=1)
raw_test_feats.shape

test_df["img_features"] = raw_test_feats.tolist()
repo.save("test_features_final", test_df)

test_feats = repo.load("test_features_final")
test_feats.head(5)

# test_generator.reset()
# raw_test_probs = model.predict(test_generator, steps=NUM_TEST_BATCHES, verbose=1)
# raw_test_probs.shape

# test_feats["img_probs"] = raw_test_probs.tolist()
# test_feats.head(5)

# repo.save("test_features", test_feats)

# Group features into clips

train_feats = repo.load("train_features_final")
val_feats = repo.load("val_features_final")
test_feats = repo.load("test_features_final")

train_feats.head(5)

val_feats.head(5)

test_feats.head(5)

def agg_features(series):
    series = sorted(series, key=lambda t: t[0])
    
    return [seq_feats[1] for seq_feats in series]

def agg_probs(series):
    series = sorted(series, key=lambda t: t[0])
    
    return [seq_probs[1] for seq_probs in series]

def make_clips(df):
    df["seq_features"] = list(zip(df.seq_num, df.img_features))
    #df["seq_probs"] = list(zip(df.seq_num, df.img_probs))
    df_clips = df\
        .groupby(by=["patient_id", "scan_id", "label", "n_slice", "clip_num"])\
        .agg({
            'file': ["min", "max"],
            'seq_features': [agg_features]#,
            #'seq_probs': [agg_probs]
        })\
        .reset_index()

    # Takes care of the multi-index after the groupby
    df_clips.columns = df_clips.columns.map(lambda t: '_'.join(t).rstrip("_"))

    df_clips = df_clips.rename(columns={
        "seq_features_agg_features": "seq_features",#"seq_probs_agg_probs": "seq_probs",
        "file_min": "clip_start_file",
        "file_max": "clip_end_file",
    })
    
    return df_clips

# train_feats = repo.load("train_features")
train_clips = make_clips(train_feats)
train_clips.head()

train_clips.shape

repo.save("train_clips_final", train_clips)

#val_feats = repo.load("val_features")
val_clips = make_clips(val_feats)
val_clips.head()

val_clips.shape

repo.save("val_clips_final", val_clips)

# test_feats = repo.load("test_features_final")
test_clips = make_clips(test_feats)
test_clips.head()

test_clips.shape

repo.save("test_clips_final", test_clips)
